pyproject.toml
[project]
name = "smart-pdf-qa"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "faiss-cpu>=1.10.0",
    "langchain>=0.3.22",
    "langchain-community>=0.3.20",
    "langchain-openai>=0.3.12",
    "langchainhub>=0.1.21",
    "openai>=1.70.0",
    "pymupdf>=1.25.5",
    "pypdf>=5.4.0",
    "python-dotenv>=1.1.0",
    "streamlit>=1.44.1",
    "tiktoken>=0.9.0",
    "watchdog>=6.0.0",
]
